relocate abstract fragments out of their base package
