+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/FragmentComponent.kt
@@ -2,8 +2,8 @@ package com.lambdaschool.hackathon_portal.di
 
 import androidx.fragment.app.Fragment
 import com.lambdaschool.hackathon_portal.ui.fragments.account.AccountFragment
-import com.lambdaschool.hackathon_portal.ui.fragments.base.BaseFragment
-import com.lambdaschool.hackathon_portal.ui.fragments.base.NavDrawerFragment
+import com.lambdaschool.hackathon_portal.ui.fragments.BaseFragment
+import com.lambdaschool.hackathon_portal.ui.fragments.NavDrawerFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.create.CreateHackathonFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.dashboard.DashboardFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.detail.AdminFragment
